dataops/requirements.txt
attrs==21.4.0
certifi==2021.10.8
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.21
munch==2.5.0
psycopg2-binary==2.9.3
six==1.16.0
